Refactor state components to use default exports
